dashboard/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
authlib>=1.3.0
httpx>=0.27.0
slowapi>=0.1.9
python-multipart>=0.0.9
aiofiles>=24.0
itsdangerous>=2.1.0
